# docs/source/API/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Reachy Mini REST API",
    "version": "0.1.0",
    "description": "HTTP and WebSocket API exposed by the Reachy Mini daemon. Interactive Swagger docs are available at `/docs` when the daemon is running."
  },
  "paths": {
...
    "/api/media/play_sound": {
      "post": {
        "summary": "Play Sound",
        "description": "Play a sound file on the robot's speaker.\n\nThe *file* field can be:\n- An absolute path on the daemon's filesystem.\n- A filename relative to the built-in assets directory.\n- A filename previously uploaded to the sounds temp directory.",
        "operationId": "play_sound_api_media_play_sound_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaySoundRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Play Sound Api Media Play Sound Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/media/sounds/upload": {
      "post": {
        "summary": "Upload Sound",
        "description": "Upload a sound file to the daemon's temporary sound directory.\n\nThe file is saved to ``/tmp/reachy_mini_sounds/<original_filename>``.\nIf a file with the same name already exists it is overwritten.\n\nThe upload is restricted to known audio extensions, capped at\n``MAX_SOUND_UPLOAD_BYTES``, and validated by content before being stored,\nso non-audio payloads cannot be written to disk.\n\nReturns:\n    JSON with the absolute *path* of the saved file on the daemon.",
        "operationId": "upload_sound_api_media_sounds_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_sound_api_media_sounds_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Upload Sound Api Media Sounds Upload Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/media/sounds/{filename}": {
      "delete": {
        "summary": "Delete Sound",
        "description": "Delete a sound file from the daemon's temporary sound directory.\n\nOnly files inside the temp directory can be deleted (no path traversal).",
        "operationId": "delete_sound_api_media_sounds__filename__delete",
        "parameters": [
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Filename"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Delete Sound Api Media Sounds  Filename  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_sound_api_media_sounds_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_sound_api_media_sounds_upload_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlaySoundRequest": {
        "properties": {
          "file": {
            "type": "string",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "PlaySoundRequest",
        "description": "Request body for the play_sound endpoint."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}